+++ lithops/executors.py
@@ -18,21 +18,29 @@
 import copy
 import signal
 import logging
+import atexit
+import os
+import pickle
+import tempfile
+import sys
+import subprocess as sp
 from functools import partial
+
 from lithops.invokers import ServerlessInvoker, StandaloneInvoker
 from lithops.storage import InternalStorage
-from lithops.storage.utils import delete_cloudobject
 from lithops.wait import wait_storage, wait_rabbitmq, ALL_COMPLETED
-from lithops.job import create_map_job, create_reduce_job, clean_job
+from lithops.job import create_map_job, create_reduce_job
 from lithops.config import default_config, extract_storage_config, \
     default_logging_config, extract_localhost_config, \
     extract_standalone_config, extract_serverless_config, LOCALHOST,\
-    SERVERLESS, STANDALONE
+    SERVERLESS, STANDALONE, CLEANER_DIR, CLEANER_LOG_FILE
 from lithops.utils import timeout_handler, is_notebook, \
     is_unix_system, is_lithops_worker, create_executor_id
 from lithops.localhost.localhost import LocalhostHandler
 from lithops.standalone.standalone import StandaloneHandler
 from lithops.serverless.serverless import ServerlessHandler
+from lithops.storage.utils import create_job_key
+
 
 logger = logging.getLogger(__name__)
 
@@ -89,6 +97,11 @@ class FunctionExecutor:
         self.executor_id = create_executor_id()
 
         self.data_cleaner = self.config['lithops'].get('data_cleaner', True)
+        if self.data_cleaner and not self.is_lithops_worker:
+            spawn_cleaner = int(self.executor_id.split('-')[1]) == 0
+            atexit.register(self.clean, spawn_cleaner=spawn_cleaner,
+                            clean_cloudobjects=False)
+
         self.rabbitmq_monitor = self.config['lithops'].get('rabbitmq_monitor', False)
 
         if self.rabbitmq_monitor:
@@ -104,7 +117,6 @@ class FunctionExecutor:
 
         self.futures = []
         self.total_jobs = 0
-        self.cleaned_jobs = set()
         self.last_call = None
 
         if mode == LOCALHOST:
@@ -418,7 +430,7 @@ class FunctionExecutor:
                 if not is_notebook():
                     print()
             if self.data_cleaner and not self.is_lithops_worker:
-                self.clean(cloudobjects=False, force=False)
+                self.clean(clean_cloudobjects=False)
             if not fs and error and is_notebook():
                 del self.futures[len(self.futures)-len(futures):]
 
@@ -502,43 +514,47 @@ class FunctionExecutor:
         create_timeline(ftrs_to_plot, dst)
         create_histogram(ftrs_to_plot, dst)
 
-    def clean(self, fs=None, cs=None, cloudobjects=True, force=True):
+    def clean(self, fs=None, cs=None, clean_cloudobjects=True, spawn_cleaner=True):
         """
-        Deletes all the files from COS. These files include the function,
-        the data serialization and the function invocation results.
+        Deletes all the temp files from storage. These files include the function,
+        the data serialization and the function invocation results. It can also clean
+        cloudobjects.
+
+        :param fs: list of futures to clean
+        :param cs: list of cloudobjects to clean
+        :param clean_cloudobjects: true/false
+        :param spawn_cleaner true/false
         """
+        os.makedirs(CLEANER_DIR, exist_ok=True)
+
+        def save_data_to_clean(data):
+            with tempfile.NamedTemporaryFile(dir=CLEANER_DIR, delete=False) as temp:
+                pickle.dump(data, temp)
 
         if cs:
-            storage_config = self.internal_storage.get_storage_config()
-            delete_cloudobject(list(cs), storage_config)
+            data = {'cos_to_clean': list(cs),
+                    'storage_config': self.internal_storage.get_storage_config()}
+            save_data_to_clean(data)
             if not fs:
                 return
 
-        futures = self.futures if not fs else fs
-        if type(futures) != list:
-            futures = [futures]
+        futures = fs or self.futures
+        futures = [futures] if type(futures) != list else futures
 
-        if not futures:
-            logger.debug('ExecutorID {} - No jobs to clean'
-                         .format(self.executor_id))
-            return
-
-        if fs or force:
-            present_jobs = {(f.executor_id, f.job_id) for f in futures
-                            if f.executor_id.count('-') == 1}
-            jobs_to_clean = present_jobs
-        else:
-            present_jobs = {(f.executor_id, f.job_id) for f in futures
-                            if f.done and f.executor_id.count('-') == 1}
-            jobs_to_clean = present_jobs - self.cleaned_jobs
-
-        if jobs_to_clean:
+        if futures:
             logger.info("ExecutorID {} - Cleaning temporary data"
                         .format(self.executor_id))
-            storage_config = self.internal_storage.get_storage_config()
-            clean_job(jobs_to_clean, storage_config,
-                      clean_cloudobjects=cloudobjects)
-            self.cleaned_jobs.update(jobs_to_clean)
+            jobs_to_clean = {create_job_key(f.executor_id, f.job_id) for f in futures
+                             if f.done and f.executor_id.count('-') == 1}
+            data = {'jobs_to_clean': jobs_to_clean,
+                    'clean_cloudobjects': clean_cloudobjects,
+                    'storage_config': self.internal_storage.get_storage_config()}
+            save_data_to_clean(data)
+
+        if (futures or cs) and spawn_cleaner:
+            log_file = open(CLEANER_LOG_FILE, 'a')
+            cmdstr = '{} -m lithops.util.cleaner'.format(sys.executable)
+            sp.Popen(cmdstr, shell=True, stdout=log_file, stderr=log_file)
 
     def dismantle(self):
         self.compute_handler.dismantle()
@@ -548,8 +564,6 @@ class FunctionExecutor:
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.invoker.stop()
-        if self.data_cleaner:
-            self.clean(force=False)
 
 
 class LocalhostExecutor(FunctionExecutor):