+++ lithops/serverless/backends/gcp_functions/entry_point.py
@@ -20,18 +20,21 @@ import base64
 import os
 import uuid
 from lithops.version import __version__
-from lithops.config import cloud_logging_config
+from lithops.config import default_logging_config
 from lithops.worker import function_handler
 from lithops.worker import function_invoker
 
-cloud_logging_config(logging.INFO)
 logger = logging.getLogger('__main__')
 
 
 def main(event, context):
     logger.info("Starting GCP Functions function execution")
     args = json.loads(base64.b64decode(event['data']).decode('utf-8'))
-    os.environ['__PW_ACTIVATION_ID'] = uuid.uuid4().hex
+    os.environ['__LITHOPS_ACTIVATION_ID'] = uuid.uuid4().hex
+
+    log_level = args['log_level']
+    default_logging_config(log_level)
+
     if 'remote_invoker' in args:
         logger.info("Lithops v{} - Starting invoker".format(__version__))
         function_invoker(args)