+++ lithops/standalone/standalone.py
@@ -26,6 +26,7 @@ from threading import Thread
 from lithops.utils import is_lithops_worker
 from lithops.serverless.utils import create_function_handler_zip
 from lithops.config import LOGS_DIR, REMOTE_INSTALL_DIR, FN_LOG_FILE
+from lithops.storage.utils import create_job_key
 
 logger = logging.getLogger(__name__)
 FH_ZIP_LOCATION = os.path.join(os.getcwd(), 'lithops_standalone.zip')
@@ -157,34 +158,40 @@ class StandaloneHandler:
         Starts a process that polls the remote log into a local file
         """
 
-        exec_id = '-'.join([executor_id, job_id])
+        job_key = create_job_key(executor_id, job_id)
 
         def log_monitor():
             os.makedirs(LOGS_DIR, exist_ok=True)
-            log_file = os.path.join(LOGS_DIR, exec_id+'.log')
-            fdout_0 = open(log_file, 'w')
-            fdout_1 = open(FN_LOG_FILE, 'a')
+            log_file = os.path.join(LOGS_DIR, job_key+'.log')
+            fdout_0 = open(log_file, 'wb')
+            fdout_1 = open(FN_LOG_FILE, 'ab')
 
             ssh_client = self.ssh_client.create_client(self.ip_address)
-            cmd = 'tail -n +1 -F /tmp/lithops/logs/{}.log'.format(exec_id)
+            cmd = 'tail -n +1 -F /tmp/lithops/logs/{}.log'.format(job_key)
             stdin, stdout, stderr = ssh_client.exec_command(cmd)
             channel = stdout.channel
             stdin.close()
             channel.shutdown_write()
 
+            data = None
             while not channel.closed:
-                readq, _, _ = select.select([channel], [], [], 10)
-                if readq and readq[0].recv_ready():
-                    data = channel.recv(len(readq[0].in_buffer)).decode()
-                    fdout_0.write(data)
-                    fdout_0.flush()
-                    fdout_1.write(data)
-                    fdout_1.flush()
-                else:
-                    cmd = 'ls /tmp/lithops/jobs/{}.done'.format(exec_id)
-                    _, out, _ = ssh_client.exec_command(cmd)
-                    if out.read().decode().strip():
-                        break
+                try:
+                    readq, _, _ = select.select([channel], [], [], 10)
+                    if readq and readq[0].recv_ready():
+                        data = channel.recv(len(readq[0].in_buffer))
+                        fdout_0.write(data)
+                        fdout_0.flush()
+                        fdout_1.write(data)
+                        fdout_1.flush()
+                    else:
+                        if data:
+                            cmd = 'ls /tmp/lithops/jobs/{}.done'.format(job_key)
+                            _, out, _ = ssh_client.exec_command(cmd)
+                            if out.read().decode().strip():
+                                break
+                        time.sleep(0.5)
+                except Exception:
+                    pass
 
         if not self.is_lithops_worker:
             Thread(target=log_monitor, daemon=True).start()
@@ -197,8 +204,8 @@ class StandaloneHandler:
         """
         executor_id = job_payload['executor_id']
         job_id = job_payload['job_id']
-        exec_id = '-'.join([executor_id, job_id])
-        log_file = os.path.join(LOGS_DIR, exec_id+'.log')
+        job_key = create_job_key(executor_id, job_id)
+        log_file = os.path.join(LOGS_DIR, job_key+'.log')
 
         if not self._is_proxy_ready():
             # The VM instance is stopped